feat(vpn): Move controls to the VPN state to a dedicated class
