+++ app/src/main/java/org/adaway/broadcast/BootReceiver.java
@@ -20,6 +20,10 @@
 
 package org.adaway.broadcast;
 
+import static android.content.Intent.ACTION_BOOT_COMPLETED;
+import static org.adaway.model.adblocking.AdBlockMethod.ROOT;
+import static org.adaway.model.adblocking.AdBlockMethod.VPN;
+
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
@@ -27,11 +31,7 @@ import android.content.Intent;
 import org.adaway.helper.PreferenceHelper;
 import org.adaway.model.adblocking.AdBlockMethod;
 import org.adaway.util.WebServerUtils;
-import org.adaway.vpn.VpnService;
-
-import static android.content.Intent.ACTION_BOOT_COMPLETED;
-import static org.adaway.model.adblocking.AdBlockMethod.ROOT;
-import static org.adaway.model.adblocking.AdBlockMethod.VPN;
+import org.adaway.vpn.VpnServiceControls;
 
 import timber.log.Timber;
 
@@ -57,7 +57,7 @@ public class BootReceiver extends BroadcastReceiver {
                     context.startActivity(prepareIntent);
                 }
                 // Start VPN service if enabled in preferences
-                VpnService.start(context);
+                VpnServiceControls.start(context);
             }
         }
     }
